feat: 生成轻量约束，程序补
